+++ addons/im_livechat/__manifest__.py
@@ -58,7 +58,7 @@ Help your customers with this chat, and analyse their feedback.
             'im_livechat/static/src/legacy/widgets/*',
             'im_livechat/static/src/legacy/widgets/*/*',
             'im_livechat/static/src/legacy/public_livechat_chatbot.js',
-            'im_livechat/static/src/legacy/website_livechat_message_chatbot.js',
+            'im_livechat/static/src/legacy/public_livechat_message_chatbot.js',
             'im_livechat/static/src/legacy/public_livechat.scss',
             'im_livechat/static/src/legacy/public_livechat_chatbot.scss',
         ],
@@ -85,7 +85,7 @@ Help your customers with this chat, and analyse their feedback.
             'im_livechat/static/src/legacy/widgets/*',
             'im_livechat/static/src/legacy/widgets/*/*',
             'im_livechat/static/src/legacy/public_livechat_chatbot.js',
-            'im_livechat/static/src/legacy/website_livechat_message_chatbot.js',
+            'im_livechat/static/src/legacy/public_livechat_message_chatbot.js',
         ],
         'web.assets_tests': [
             'im_livechat/static/tests/tours/**/*',
@@ -195,7 +195,7 @@ Help your customers with this chat, and analyse their feedback.
             'im_livechat/static/src/legacy/widgets/*',
             'im_livechat/static/src/legacy/widgets/*/*',
             'im_livechat/static/src/legacy/public_livechat_chatbot.js',
-            'im_livechat/static/src/legacy/website_livechat_message_chatbot.js',
+            'im_livechat/static/src/legacy/public_livechat_message_chatbot.js',
 
             ('include', 'web._assets_helpers'),
