+++ addons/im_livechat/controllers/main.py
@@ -28,7 +28,7 @@ class LivechatController(http.Controller):
     def load_templates(self, **kwargs):
         templates = [
             'im_livechat/static/src/legacy/widgets/feedback/feedback.xml',
-            'im_livechat/static/src/legacy/widgets/abstract_thread_window/abstract_thread_window.xml',
+            'im_livechat/static/src/legacy/widgets/public_livechat_window/public_livechat_window.xml',
             'im_livechat/static/src/legacy/widgets/thread/thread.xml',
             'im_livechat/static/src/legacy/widgets/document_viewer/document_viewer.xml',
             'im_livechat/static/src/legacy/public_livechat_chatbot.xml',
