+++ addons/im_livechat/static/src/legacy/models/public_livechat.js
