+++ addons/im_livechat/static/src/legacy/public_livechat.scss
@@ -1,5 +1,5 @@
 //-------------------------------------
-// legacy: mail/abstract_thread_window.scss
+// legacy: mail/public_livechat_window.scss
 //-------------------------------------
 
 $o-mail-thread-window-zindex: $zindex-modal + 1 !default;
