+++ addons/im_livechat/static/src/legacy/widgets/public_livechat_window/public_livechat_window.xml
